inject-connect cmd: deprecate -consul-ca-cert flag
Now that the inject-connect command has Consul's HTTP flags,
we should use only one flag to control TLS communications with the local agents.

Currently, we set both flags. The -consul-ca-cert flag sets the
CA cert for the init container, the envoy proxy sidecar, and the lifecycle
sidecar, and they use that CA to talk to their local agent.

The -ca-file flag controls the CA cert that is used by the handler
to talk to its local agent (this happens only if namespaces are enabled).

Although these agents could be on different hosts,
the CA that singed their certificates must be the same,
so there is no need to have two different flags.